+++ pkg/utl/convert/convert_test.go
@@ -4,9 +4,6 @@ import (
 	"reflect"
 	"testing"
 
-	graphql "go-template/gqlmodels"
-	"go-template/models"
-
 	"github.com/volatiletech/null/v8"
 )
 
@@ -256,38 +253,6 @@ func TestPointerStringToNullDotInt(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := PointerStringToNullDotInt(tt.args.s); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("PointerStringToNullDotInt() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestUsersToGraphQlUsers(t *testing.T) {
-	type args struct {
-		u models.UserSlice
-	}
-	tests := []struct {
-		name string
-		args args
-		want []*graphql.User
-	}{
-		{
-			name: "Success",
-			args: args{
-				u: models.UserSlice{{
-					ID: 1,
-				}},
-			},
-			want: []*graphql.User{
-				{
-					ID: "1",
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := UsersToGraphQlUsers(tt.args.u, 1); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("UsersToGraphQlUsers() = %v, want %v", got, tt.want)
 			}
 		})
 	}
